HELIO-2302 Upgrade to Rails 5.2+
