--- config/initializers/actionview_monkeypatch.rb
@@ -1,29 +1,0 @@
-# frozen_string_literal: true
-
-# See HELIO-3304
-# After we update to rails >= 5.2.4.2 this can go away.
-
-# rubocop:disable Performance/RedundantMerge, Rails/OutputSafety
-ActionView::Helpers::JavaScriptHelper::JS_ESCAPE_MAP.merge!(
-  {
-    "`" => "\\`",
-    "$" => "\\$"
-  }
-)
-
-module ActionView::Helpers::JavaScriptHelper
-  alias :old_ej :escape_javascript
-  alias :old_j :j
-
-  def escape_javascript(javascript)
-    javascript = javascript.to_s
-    if javascript.empty?
-      result = ""
-    else
-      result = javascript.gsub(/(\\|<\/|\r\n|\342\200\250|\342\200\251|[\n\r"']|[`]|[$])/u, JS_ESCAPE_MAP)
-    end
-    javascript.html_safe? ? result.html_safe : result
-  end
-
-  alias :j :escape_javascript
-end
